+++ changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
-**Test Status**
-* Frontend API testing is largely requesting the same endpoints but from an end-user perspective.
-* Hence, most of the tests already pass. (69 / 87)
+**./routes/nodes.js**
+* New file - Contains Node API endpoints.
+
+---
+
+**./server.js**
+* Added Node API endpoints.